+++ migrator/database_connector.py
@@ -462,10 +462,25 @@ class DatabaseConnector(ABC):
         { ordinary_number: {
             'domain_schema': schema_name,
             'domain_name': domain_name,
-            'domain_sql': domain_sql,
+            'source_domain_sql':
+                - Original SQL statement to create the domain from the source database
+                - Contains all the specific syntax of the source database
+            'standardized_domain_sql':
+                - SQL statement to create the domain stript of all the specific syntax of the source database
+                - Should contains only the standard SQL syntax and only the conditions
             'domain_comment': domain_comment
             }
         }
+        """
+        pass
+
+    @abstractmethod
+    def get_create_domain_sql(self, settings):
+        """
+        This function is currently relevant only for target database
+        Centralizes creation of SQL DDL statement for domains
+        settings:
+            -
         """
         pass