+++ migrator/informix_connector.py
@@ -40,7 +40,6 @@ class InformixConnector(DatabaseConnector):
         else:
             raise ValueError(f"Unsupported connectivity: {self.config_parser.get_connectivity(self.source_or_target)}")
 
-
     def disconnect(self):
         try:
             if self.connection:
@@ -155,13 +154,6 @@ class InformixConnector(DatabaseConnector):
                     'column_default': re.sub(r'[^\x20-\x7E]', ' ', row[5]).strip() if row[5] else '',
                     'column_comment': ''
                 }
-
-                # # if self.config_parser.get_log_level() == 'DEBUG':
-                # #     self.logger.debug(f"0 default: {result[row[0]]['column_default']}")
-                # # checking for default values substitution with the origingal data type
-                # if migrator_tables:
-                #     if result[row[0]]['column_default'] != '':
-                #         result[row[0]]['column_default'] = migrator_tables.check_default_values_substitution(result[row[0]]['name'], result[row[0]]['type'], result[row[0]]['column_default'])
 
             cursor.close()
             self.disconnect()
@@ -1046,6 +1038,7 @@ class InformixConnector(DatabaseConnector):
 
     def migrate_table(self, migrate_target_connection, settings):
         part_name = 'initialize'
+        target_table_rows = 0
         try:
             worker_id = settings['worker_id']
             source_schema = settings['source_schema']
@@ -1055,19 +1048,31 @@ class InformixConnector(DatabaseConnector):
             target_schema = settings['target_schema']
             target_table = settings['target_table']
             target_columns = settings['target_columns']
-            primary_key_columns = settings['primary_key_columns']
+            # primary_key_columns = settings['primary_key_columns']
             batch_size = settings['batch_size']
             migrator_tables = settings['migrator_tables']
-            source_table_rows = self.get_rows_count(source_schema, source_table)
             migration_limitation = settings['migration_limitation']
+
+            source_table_rows = self.get_rows_count(source_schema, source_table)
+
+            ## source_schema, source_table, source_table_id, source_table_rows, worker_id, target_schema, target_table, target_table_rows
+            migrator_tables_settings = {
+                'worker_id': worker_id,
+                'source_table_id': source_table_id,
+                'source_schema': source_schema,
+                'source_table': source_table,
+                'target_schema': target_schema,
+                'target_table': target_table,
+                'source_table_rows': source_table_rows,
+                'target_table_rows': target_table_rows,
+            }
+            protocol_id = migrator_tables.insert_data_migration(migrator_tables_settings)
 
             if source_table_rows == 0:
                 self.logger.info(f"Worker {worker_id}: Table {source_table} is empty - skipping data migration.")
-                migrator_tables.insert_data_migration(source_schema, source_table, source_table_id, source_table_rows, worker_id, target_schema, target_table, 0)
                 return 0
             else:
                 self.logger.info(f"Worker {worker_id}: Table {source_table} has {source_table_rows} rows - starting data migration.")
-                protocol_id = migrator_tables.insert_data_migration(source_schema, source_table, source_table_id, source_table_rows, worker_id, target_schema, target_table, 0)
                 # Fetch the data in batches
                 # Open a cursor and fetch rows in batches
                 query = f'''SELECT * FROM "{source_schema}".{source_table}'''
@@ -1581,6 +1586,10 @@ class InformixConnector(DatabaseConnector):
     def fetch_domains(self, schema: str):
         # Placeholder for fetching domains
         return {}
+
+    def get_create_domain_sql(self, settings):
+        # Placeholder for generating CREATE DOMAIN SQL
+        return ""
 
     def testing_select(self):
         return "SELECT 1"